refactor: use uintptr_t for memory addresses throughout
Replace uint32_t with uintptr_t for all variables and parameters that
hold memory addresses rather than game data values. This covers function
pointer params (GetObjectPointer return, unit/pointer params, FUN_ONRIGHTCLICK),
CandidateInfo::pointer, and all ReadMemory template arguments used to
dereference pointer-valued struct fields. Game data types (guid, id,
flags, health) are unchanged.

+++ include/Game.h
@@ -16,7 +16,7 @@ enum class ObjectType : uint32_t {
   CORPSE
 };
 
-typedef void(__thiscall *FUN_ONRIGHTCLICK)(uint32_t pointer, int autoloot);
+typedef void(__thiscall *FUN_ONRIGHTCLICK)(uintptr_t pointer, int autoloot);
 typedef void(__fastcall *FrameScript_RegisterFunction)(const char *, uintptr_t);
 typedef void(__cdecl *lua_error)(void *, const char *);
 typedef bool(__fastcall *lua_isnumber)(void *, int);
@@ -32,18 +32,18 @@ namespace Game {
 auto const RegisterFunction =
     (FrameScript_RegisterFunction)Offsets::FUN_REGISTER_LUA_FUNCTION;
 
-uint32_t __stdcall GetObjectPointer(uint64_t guid);
+uintptr_t __stdcall GetObjectPointer(uint64_t guid);
 
-C3Vector GetObjectPosition(uint32_t pointer);
+C3Vector GetObjectPosition(uintptr_t pointer);
 
-int GetUnitHealth(uint32_t unit);
+int GetUnitHealth(uintptr_t unit);
 
-C3Vector GetUnitPosition(uint32_t unit);
+C3Vector GetUnitPosition(uintptr_t unit);
 
-void Interact(uint32_t pointer, int autoloot, uintptr_t fun_ptr);
+void Interact(uintptr_t pointer, int autoloot, uintptr_t fun_ptr);
 
-bool IsUnitLootable(uint32_t unit);
-bool IsUnitSkinnable(uint32_t unit);
+bool IsUnitLootable(uintptr_t unit);
+bool IsUnitSkinnable(uintptr_t unit);
 
 void SetTarget(uint64_t guid);
